sip-build: Cross platform build support
Projects may use the PyQtBuilder in pyproject.toml to automatically
generate the build files and execute the build.

PyQtBuilder/builder.py invokes buildable.get_module_extension() to
create rules in QMake project files for building Python extensions
for the target.

However, this determines the module extension using the importlib module.
This provides the extension suffix used for Python modules that
may be imported. However, this is incorrect when running Python
in a cross-build environment.  For OpenEmbedded, the correct
extension suffix for the target may be obtained by referring to
the sysconfig EXT_SUFFIX configuration variable.

Upstream-Status: Inappropriate [oe build specific]

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>